Fix spawning particles on startup

+++ ruvu_mcl/src/config.hpp
@@ -31,7 +31,7 @@ struct Config
   Config(const ruvu_mcl::AMCLConfig & config);
   Config() = default;
 
-  int max_particles;
+  size_t max_particles;
   double update_min_d;
   double update_min_a;
   int resample_interval;
